refactor dpm mandate analysis routes
